Rename Destroy/Store requests to Delete*/Create*
