+++ src/Http/Controllers/Web/Bulk/PostController.php
@@ -4,12 +4,12 @@ namespace TeamTeaTime\Forum\Http\Controllers\Web\Bulk;
 
 use Illuminate\Http\RedirectResponse;
 use TeamTeaTime\Forum\Http\Controllers\Web\BaseController;
-use TeamTeaTime\Forum\Http\Requests\Bulk\DestroyPosts;
+use TeamTeaTime\Forum\Http\Requests\Bulk\DeletePosts;
 use TeamTeaTime\Forum\Http\Requests\Bulk\RestorePosts;
 
 class PostController extends BaseController
 {
-    public function destroy(DestroyPosts $request): RedirectResponse
+    public function destroy(DeletePosts $request): RedirectResponse
     {
         $count = $request->fulfill()->count();
