+++ src/Http/Controllers/Web/ThreadController.php
@@ -11,6 +11,7 @@ use TeamTeaTime\Forum\Events\UserMarkingNew;
 use TeamTeaTime\Forum\Events\UserViewingRecent;
 use TeamTeaTime\Forum\Events\UserViewingThread;
 use TeamTeaTime\Forum\Events\UserViewingUnread;
+use TeamTeaTime\Forum\Http\Requests\CreateThread;
 use TeamTeaTime\Forum\Http\Requests\DestroyThread;
 use TeamTeaTime\Forum\Http\Requests\LockThread;
 use TeamTeaTime\Forum\Http\Requests\MarkThreadsAsRead;
@@ -18,7 +19,6 @@ use TeamTeaTime\Forum\Http\Requests\MoveThread;
 use TeamTeaTime\Forum\Http\Requests\PinThread;
 use TeamTeaTime\Forum\Http\Requests\RenameThread;
 use TeamTeaTime\Forum\Http\Requests\RestoreThread;
-use TeamTeaTime\Forum\Http\Requests\StoreThread;
 use TeamTeaTime\Forum\Http\Requests\UnlockThread;
 use TeamTeaTime\Forum\Http\Requests\UnpinThread;
 use TeamTeaTime\Forum\Models\Category;
@@ -112,7 +112,7 @@ class ThreadController extends BaseController
         return view('forum::thread.create', compact('category'));
     }
 
-    public function store(StoreThread $request, Category $category): RedirectResponse
+    public function store(CreateThread $request, Category $category): RedirectResponse
     {
         $thread = $request->fulfill();
