+++ src/Http/Requests/CreateThread.php
@@ -9,7 +9,7 @@ use TeamTeaTime\Forum\Interfaces\FulfillableRequest;
 use TeamTeaTime\Forum\Models\Category;
 use TeamTeaTime\Forum\Models\Thread;
 
-class StoreThread extends FormRequest implements FulfillableRequest
+class CreateThread extends FormRequest implements FulfillableRequest
 {
     public function authorize(): bool
     {
