+++ src/Http/Requests/DeletePost.php
@@ -9,7 +9,7 @@ use TeamTeaTime\Forum\Http\Requests\Traits\HandlesDeletion;
 use TeamTeaTime\Forum\Interfaces\FulfillableRequest;
 use TeamTeaTime\Forum\Models\Post;
 
-class DestroyPost extends FormRequest implements FulfillableRequest
+class DeletePost extends FormRequest implements FulfillableRequest
 {
     use HandlesDeletion;
